Revert back to 4030b73 and fix the virtual function wrapping (#352)

+++ include/galsim/SBBoxImpl.h
@@ -79,9 +79,9 @@ namespace galsim {
         { return fillXImage(I,gain); }
         double doFillXImage(ImageView<double>& I, double gain) const
         { return fillXImage(I,gain); }
-        double doFillXImage(ImageView<short>& I, double gain) const
+        double doFillXImage(ImageView<int32_t>& I, double gain) const
         { return fillXImage(I,gain); }
-        double doFillXImage(ImageView<int>& I, double gain) const
+        double doFillXImage(ImageView<int16_t>& I, double gain) const
         { return fillXImage(I,gain); }
 
     private:
